ядро с лонгами и датами
git-svn-id: file:///var/www/svn2/oecdo/oecde@597 50b15a66-aedb-45cf-97b2-84293224aabc

+++ core/src/main/java/org/sgu/oecde/core/BasicDao.java
@@ -50,7 +50,7 @@ public class BasicDao<T extends BasicItem> extends HibernateDaoSupport implement
      * {@inheritDoc}
      */
     @Override
-    public T getById(final int id) throws DataAccessException{
+    public T getById(final Long id) throws DataAccessException{
         T item = (T) getSession().get(type, id);
         return item;
     }
